modify courses, add video features and promo code features

+++ app/Services/CourseService.php
@@ -1,12 +1,21 @@
 <?php
 
 namespace App\Services;
+use App\DTO\ApproveCourseDto;
+use App\Helper\videoHelper;
+use App\Jobs\ApproveScheduledCourseJob;
 use App\Models\Course;
 use App\Dto\CourseDto;
+use App\Models\CourseRequirement;
+use App\Models\Video;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Facades\DB;
 class CourseService
 {
+    public $video_helper;
+    public function __construct(videoHelper $video_helper){$this->video_helper = $video_helper;}
+
     public function getAllActive()//done
     {
         try {
@@ -26,6 +35,23 @@ class CourseService
     {
         try {
             $course = Course::find($id);
+            $requirements = CourseRequirement::where('course_id', $course->id)->pluck('requirements');
+            $course['requirements'] = $requirements;
+
+            $videos = Video::query()
+                ->where('course_id',$course->id)
+                ->get();
+            $courseTotalDurationSeconds = 0;
+
+            foreach ($videos as $video){
+                $courseTotalDurationSeconds += $this->video_helper->hmsToSeconds($video->duration);
+            }
+
+            $course['total_duration'] = $this->video_helper->secondsToHms($courseTotalDurationSeconds);
+            $course['videos_count'] = $videos->count();
+            $user_count = $course->user()->count();
+            $course['users_count'] = $user_count - 1;
+
             if (!$course) {
                 return ['data' => null, 'message' => 'Course not found'];
             }
@@ -39,13 +65,7 @@ class CourseService
     public function getMyCourses()
     {
         try {
-            $user = auth()->user();
-            if (!$user || auth()->user()->role !== 'teacher'){
-                return ['data' => null, 'message' => 'Unauthorized - teacher only'];
-
-            }
-//            $userId = auth()->id();
-            $courses = Course::where('user_id', $user->auth()->id )->get();
+            $courses = Course::where('user_id', Auth::id() )->get();
             return ['data' => $courses, 'message' => 'Your courses retrieved successfully'];
         } catch (\Exception $e) {
             Log::error('Fetching user courses failed', ['error' => $e->getMessage()]);
@@ -55,9 +75,15 @@ class CourseService
 
     public function getEndedCourses()//done
     {
+        $user = Auth::user();
+        if (!$user->hasRole('supervisor')){
+            return [
+                'data' => null,
+                'message' => 'must be a teacher to update a course requirement'
+            ];
+        }
         try {
-            $today = now()->toDateString();
-            $courses = Course::where('end_date', '<', $today)->where('status', 'published')->get();
+            $courses = Course::where('end_date', '<', now()->toDateString())->get();
             if($courses->isempty()) {
                 return ['data' => null, 'message' => 'there is not  courses today '];
             }else{
@@ -71,13 +97,26 @@ class CourseService
 
     public function delete($id)
     {
+        $user = Auth::user();
+        if (!$user->hasRole(['teacher','supervisor'])){
+            return [
+                'data' => null,
+                'message' => 'must be a supervisor or teacher to delete a course'
+            ];
+        }
         DB::beginTransaction();
         try {
             $course = Course::find($id);
             if (!$course) {
-                DB::rollBack();
                 return ['data' => null, 'message' => 'Course not found'];
             }
+            $course->user()->detach();
+            $course->rate()->delete();
+            $course->requirements()->delete();
+            $course->exams()->delete();
+            $course->purchases()->delete();
+            $course->videos()->delete();
+
             $course->delete();
             DB::commit();
             Log::info('Course deleted', ['id' => $id]);
@@ -91,9 +130,18 @@ class CourseService
 
     public function store(CourseDto $dto)
     {
+        $user = Auth::user();
+        if (!$user->hasRole('teacher')){
+            return [
+                'data' => null,
+                'message' => 'must be a teacher to create a course'
+            ];
+        }
         DB::beginTransaction();
         try {
-            $course = Course::create([
+            //update this function to only get (array)$dto instead of this long code
+            $course = Course::query()
+                ->create([
                 'course_name' => $dto->course_name,
                 'description' => $dto->description,
                 'rating' => $dto->rating,
@@ -101,12 +149,30 @@ class CourseService
                 'is_paid' => $dto->is_paid,
                 'start_date' => $dto->start_date,
                 'end_date' => $dto->end_date,
-                'user_id' => auth()->id(),
+                'user_id' => Auth::id(),
                 'category_id' => $dto->category_id,
             ]);
+            //to set up the requirements of this course
+            $requirements = request()->input('requirements', []); //array of strings
+            foreach ($requirements as $req) {
+                CourseRequirement::create([
+                    'course_id' => $course->id,
+                    'requirements' => $req,
+                ]);
+            }
+            //create a user_courses record
+            $course->user()->attach(Auth::id(),[
+                'is_completed' => false,
+                'certificate_id' => null,
+            ]);
+            $course['requirements'] = $requirements;
             DB::commit();
             Log::info('Course created', ['id' => $course->id]);
-            return ['data' => $course, 'message' => 'Course created successfully'];
+
+            return [
+                'data' => $course,
+                'message' => 'Course created successfully, wait for approval'
+            ];
         } catch (\Exception $e) {
             DB::rollBack();
             Log::error('Course creation failed', ['error' => $e->getMessage()]);
@@ -114,27 +180,23 @@ class CourseService
         }
     }
 
-    public function update($id, CourseDto $dto)
+    public function update($id, CourseDto $dto): array
     {
+        $user = Auth::user();
+        if (!$user->hasRole('teacher')){
+            return [
+                'data' => null,
+                'message' => 'must be a teacher to create a course'
+            ];
+        }
         DB::beginTransaction();
         try {
             $course = Course::find($id);
             if (!$course) {
-                DB::rollBack();
                 return ['data' => null, 'message' => 'Course not found'];
             }
 
-            $course->update([
-                'course_name' => $dto->course_name,
-                'description' => $dto->description,
-                'rating' => $dto->rating,
-                'status' => $dto->status,
-                'is_paid' => $dto->is_paid,
-                'start_date' => $dto->start_date,
-                'end_date' => $dto->end_date,
-                'category_id' => $dto->category_id,
-            ]);
-
+            $course->update((array)$dto);
             DB::commit();
             Log::info('Course updated', ['id' => $id]);
             return ['data' => $course, 'message' => 'Course updated successfully'];
@@ -142,6 +204,130 @@ class CourseService
             DB::rollBack();
             Log::error('Course update failed', ['error' => $e->getMessage(), 'id' => $id]);
             return ['data' => null, 'message' => 'Failed to update course'];
+        }
+    }
+
+    public function get_pending_courses(): array
+    {
+        $user = Auth::user();
+        if (!$user->hasRole('supervisor')){
+            return [
+                'data' => null,
+                'message' => 'must be a teacher to update a course requirement'
+            ];
+        }
+        try {
+            $courses = Course::where('status', '=', 'pending_approval')->get();
+            if($courses->isempty()) {
+                return ['data' => null, 'message' => 'there is not  courses right now'];
+            }else{
+                return ['data' => $courses, 'message' => 'pending courses retrieved successfully'];
+            }
+        } catch (\Exception $e) {
+            Log::error('Fetching pending courses failed', ['error' => $e->getMessage()]);
+            return ['data' => null, 'message' => 'Failed to fetch pending courses'];
+        }
+    }
+
+    public function update_course_requirements($request,$requirement_id): array
+    {
+        $user = Auth::user();
+        if (!$user->hasRole('teacher')){
+            return [
+                'data' => null,
+                'message' => 'must be a teacher to update a course requirement'
+            ];
+        }
+        DB::beginTransaction();
+        try {
+            $requirement = CourseRequirement::find($requirement_id);
+            if (!$requirement) {
+                return ['data' => null, 'message' => 'requirement not found'];
+            }
+
+            $requirement->update(['requirements' => $request->requirements]);
+            DB::commit();
+            Log::info('requirement updated', ['id' => $requirement_id]);
+            return ['data' => $requirement, 'message' => 'requirement updated successfully'];
+        } catch (\Exception $e) {
+            DB::rollBack();
+            Log::error('requirement update failed', ['error' => $e->getMessage(), 'id' => $requirement_id]);
+            return ['data' => null, 'message' => 'Failed to update requirement'];
+        }
+    }
+
+    public function delete_course_requirements($requirement_id): array
+    {
+        $user = Auth::user();
+        if (!$user->hasRole('teacher')){
+            return [
+                'data' => null,
+                'message' => 'must be a teacher to delete a course requirement'
+            ];
+        }
+        DB::beginTransaction();
+        try {
+            $requirement = CourseRequirement::find($requirement_id);
+            if (!$requirement) {
+                return ['data' => null, 'message' => 'requirement not found'];
+            }
+
+            $requirement->delete();
+            DB::commit();
+            Log::info('requirement deleted', ['id' => $requirement_id]);
+            return ['data' => null, 'message' => 'requirement deleted successfully'];
+        } catch (\Exception $e) {
+            DB::rollBack();
+            Log::error('requirement delete failed', ['error' => $e->getMessage(), 'id' => $requirement_id]);
+            return ['data' => null, 'message' => 'Failed to delete requirement'];
+        }
+    }
+
+    public function approve_update_course_status(ApproveCourseDto $dto,$course_id): array
+    {
+        $user = Auth::user();
+        if (!$user->hasRole('supervisor')){
+            return [
+                'data' => null,
+                'message' => 'must be a supervisor to update a course requirement'
+            ];
+        }
+        try {
+            $course = Course::query()->find($course_id);
+            if (!$course) {
+                return ['data' => null, 'message' => 'course not found'];
+            }
+
+            if ($dto->status === 'approved') {
+
+                //if the start date is in  the future
+                if ($course->start_date && now()->lt($course->start_date)) {
+                    //automatically publish the course
+                    ApproveScheduledCourseJob::dispatch($course->id)->delay($course->start_date);
+                    $course->update(['status' => 'approved']);
+
+                    Log::info('course scheduled for automatic approval', ['id' => $course_id]);
+                    return [
+                        'data' => null,
+                        'message' => 'course has been scheduled for automatic approval'
+                    ];
+                }
+                //immeditlly approve the course
+                $course->update(['status' => 'published']);
+
+                Log::info('course approved immediately', ['id' => $course_id]);
+                return [
+                    'data' => null,
+                    'message' => 'Course approved successfully'
+                ];
+            }
+            //if there is another status (rejected ...)
+            $course->update(['status' => $dto->status]);
+            Log::info('course approved', ['id' => $course_id]);
+            return ['data' => null, 'message' => 'course approved successfully'];
+        } catch (\Exception $e) {
+            Log::error('course approve failed', ['error' => $e->getMessage(), 'id' => $course_id]);
+            return ['data' => null, 'message' => 'Failed to approve course'];
         }
     }
 }
